Check Invite Already Exists
Implemented an if() statement that checks if the invite doesn't exist, or the invite is already set to DECLINED.  If so -> it will be inserted or updated appropriately.  If not, it will return null without any further steps.  This is done to prevent an additional invitation to the same project, or an update from accepted to pending for an already existing invite.